cloud-jobs/export-qdrant-to-gcs/requirements.txt
qdrant-client==1.15.1
google-cloud-storage==2.10.0
